refactor: cd 컴포넌트 위치 common으로 이동
